ASTDiff: Extract MenuBar class from MonacoDiffView
