+++ src/main/java/gui/webdiff/MonacoDiffView.java
@@ -219,61 +219,6 @@ public class MonacoDiffView implements Renderable {
             return (t.getParent() != null)
                     ? t.getParent().getType() + "/" + t.getType() + "/" + t.getPos() + "/" +  t.getEndPos() : t.getType().toString() + t.getPos() + t.getEndPos();
     }
-
-    private static class MenuBar implements Renderable {
-        private final String toolName;
-        private final int id;
-        private final int numOfDiffs;
-        private String getNextHRef(){
-            return "/monaco-diff/" + (id + 1) % numOfDiffs;
-
-        }
-        private String getPrevHRef(){
-            return "/monaco-diff/" + (id - 1) % numOfDiffs;
-        }
-
-        public MenuBar(String toolName, int id, int numOfDiffs) {
-            this.toolName = toolName;
-            this.id = id;
-            this.numOfDiffs = numOfDiffs;
-        }
-        @Override
-        public void renderOn(HtmlCanvas html) throws IOException {
-            html
-            .div(class_("col")).content("Generated by " + toolName)
-            .div(class_("col"))
-                .div(class_("btn-toolbar justify-content-end"))
-                    .div(class_("btn-group mr-2"))
-                        .button(class_("btn btn-primary btn-sm").id("legend")
-                                .add("data-bs-container", "body")
-                                .add("data-bs-toggle", "popover")
-                                .add("data-bs-placement", "bottom")
-                                .add("data-bs-html", "true")
-                                .add("data-bs-content", "<span class=&quot;deleted&quot;>&nbsp;&nbsp;</span> deleted<br>"
-                                        + "<span class=&quot;inserted&quot;>&nbsp;&nbsp;</span> inserted<br>"
-                                        + "<span class=&quot;moved&quot;>&nbsp;&nbsp;</span> moved<br>"
-                                        + "<span class=&quot;updated&quot;>&nbsp;&nbsp;</span> updated<br>", false)
-                        ).content("Legend")
-                        .button(class_("btn btn-primary btn-sm").id("shortcuts")
-                                .add("data-bs-toggle", "popover")
-                                .add("data-bs-placement", "bottom")
-                                .add("data-bs-html", "true")
-                                .add("data-bs-content", "<b>Alt + q</b> quit<br><b>Alt + l</b> list<br>"
-                                        + "<b>Alt + t</b> top<br><b>Alt + b</b> bottom <br><b>Alt + w</b> toggle word wrap", false)
-                        )
-                        .content("Shortcuts")
-                    ._div()
-                    .div(class_("btn-group"))
-                        .a(class_("btn btn-default btn-sm btn-primary").href("/list")).content("Back")
-                        .a(class_("btn btn-default btn-sm btn-primary").href(getPrevHRef())).content("prev")
-                        .a(class_("btn btn-default btn-sm btn-primary").href(getNextHRef())).content("next")
-                        .a(class_("btn btn-default btn-sm btn-danger").href("/quit")).content("Quit")
-                    ._div()
-                ._div()
-            ._div();
-        }
-    }
-
     private static class Header implements Renderable {
         @Override
         public void renderOn(HtmlCanvas html) throws IOException {
